refactor(platform): split theme/migration.rs into module (Task 5.5)
