+++ crates/katana-platform/src/theme/builder.rs
@@ -1,8 +1,6 @@
 use crate::theme::palettes::*;
 use crate::theme::preset::PresetColorData;
-use crate::theme::types::{
-    CodeColors, PreviewColors, Rgb, Rgba, SystemColors, ThemeMode,
-};
+use crate::theme::types::{CodeColors, PreviewColors, Rgb, Rgba, SystemColors, ThemeMode};
 
 pub(crate) const fn lighten(color: Rgb, amount: u8) -> Rgb {
     Rgb {
