+++ crates/katana-platform/src/theme/migration/mod.rs
@@ -1,111 +1,10 @@
-const ERROR_R: u8 = 0xFF;
-const ERROR_G: u8 = 0x55;
-const ERROR_B: u8 = 0x55;
+mod constants;
+mod legacy_types;
 
-use crate::theme::types::{
-    CodeColors, PreviewColors, Rgb, Rgba, SystemColors, ThemeColors, ThemeMode,
-};
+use crate::theme::migration::constants::*;
+use crate::theme::migration::legacy_types::{ThemeColorsDef, ThemeColorsLegacyData};
+use crate::theme::types::{CodeColors, PreviewColors, SystemColors, ThemeColors, ThemeMode};
 use serde::{Deserialize, Deserializer};
-
-// ── Migration fallback color constants ──────────────────────────
-// Default values used when converting from legacy format to the new 3-layer structure.
-// Defined in dark/light pairs for use in the "is_dark" branch.
-
-const LEGACY_SUCCESS_DARK: Rgb = Rgb {
-    r: 80,
-    g: 200,
-    b: 80,
-};
-const LEGACY_SUCCESS_LIGHT: Rgb = Rgb {
-    r: 20,
-    g: 160,
-    b: 20,
-};
-const LEGACY_BUTTON_BG_DARK: Rgba = Rgba {
-    r: 80,
-    g: 80,
-    b: 80,
-    a: 50,
-};
-const LEGACY_BUTTON_BG_LIGHT: Rgba = Rgba {
-    r: 160,
-    g: 160,
-    b: 160,
-    a: 50,
-};
-const LEGACY_BUTTON_ACTIVE_DARK: Rgba = Rgba {
-    r: 120,
-    g: 120,
-    b: 120,
-    a: 100,
-};
-const LEGACY_BUTTON_ACTIVE_LIGHT: Rgba = Rgba {
-    r: 200,
-    g: 200,
-    b: 200,
-    a: 100,
-};
-
-const LEGACY_LINE_NUMBER_DARK: Rgb = Rgb {
-    r: 100,
-    g: 100,
-    b: 100,
-};
-const LEGACY_LINE_NUMBER_LIGHT: Rgb = Rgb {
-    r: 160,
-    g: 160,
-    b: 160,
-};
-const LEGACY_CURRENT_LINE_DARK: Rgba = Rgba {
-    r: 255,
-    g: 255,
-    b: 255,
-    a: 15,
-};
-const LEGACY_CURRENT_LINE_LIGHT: Rgba = Rgba {
-    r: 0,
-    g: 0,
-    b: 0,
-    a: 15,
-};
-const LEGACY_HOVER_LINE_DARK: Rgba = Rgba {
-    r: 255,
-    g: 255,
-    b: 255,
-    a: 10,
-};
-const LEGACY_HOVER_LINE_LIGHT: Rgba = Rgba {
-    r: 0,
-    g: 0,
-    b: 0,
-    a: 10,
-};
-
-const LEGACY_DEFAULT_WARNING: Rgb = Rgb {
-    r: 255,
-    g: 140,
-    b: 0,
-};
-const LEGACY_DEFAULT_BORDER: Rgb = Rgb {
-    r: 60,
-    g: 60,
-    b: 60,
-};
-const LEGACY_DEFAULT_SELECTION: Rgb = Rgb {
-    r: 80,
-    g: 80,
-    b: 80,
-};
-const LEGACY_DEFAULT_CODE_BG: Rgb = Rgb {
-    r: 30,
-    g: 30,
-    b: 30,
-};
-const LEGACY_DEFAULT_PREVIEW_BG: Rgb = Rgb {
-    r: 35,
-    g: 35,
-    b: 35,
-};
 
 // ── Deserialize implementation ──────────────────────────────────
 
@@ -116,66 +15,6 @@ impl<'de> Deserialize<'de> for ThemeColors {
     {
         deserialize_theme_colors(deserializer)
     }
-}
-
-#[derive(Deserialize)]
-#[serde(remote = "ThemeColors")]
-struct ThemeColorsDef {
-    pub name: String,
-    pub mode: ThemeMode,
-    pub system: SystemColors,
-    pub code: CodeColors,
-    pub preview: PreviewColors,
-}
-
-#[derive(Deserialize)]
-struct ThemeColorsLegacyData {
-    pub name: String,
-    pub mode: ThemeMode,
-    pub background: Rgb,
-    pub panel_background: Rgb,
-    pub text: Rgb,
-    pub text_secondary: Rgb,
-    pub accent: Rgb,
-    pub title_bar_text: Rgb,
-    pub file_tree_text: Rgb,
-    pub active_file_highlight: Rgba,
-    #[serde(default = "default_warning_text")]
-    pub warning_text: Rgb,
-    #[serde(default = "default_error_text")]
-    pub error_text: Rgb,
-    #[serde(default = "default_border")]
-    pub border: Rgb,
-    #[serde(default = "default_selection")]
-    pub selection: Rgb,
-    #[serde(default = "default_code_background")]
-    pub code_background: Rgb,
-    #[serde(default = "default_preview_background")]
-    pub preview_background: Rgb,
-}
-
-fn default_error_text() -> Rgb {
-    Rgb {
-        r: ERROR_R,
-        g: ERROR_G,
-        b: ERROR_B,
-    }
-}
-
-fn default_warning_text() -> Rgb {
-    LEGACY_DEFAULT_WARNING
-}
-fn default_border() -> Rgb {
-    LEGACY_DEFAULT_BORDER
-}
-fn default_selection() -> Rgb {
-    LEGACY_DEFAULT_SELECTION
-}
-fn default_code_background() -> Rgb {
-    LEGACY_DEFAULT_CODE_BG
-}
-fn default_preview_background() -> Rgb {
-    LEGACY_DEFAULT_PREVIEW_BG
 }
 
 pub(crate) fn deserialize_theme_colors<'de, D>(deserializer: D) -> Result<ThemeColors, D::Error>
